+++ src/web/components/cards/MarketCard.tsx
@@ -24,7 +24,7 @@ export default function MarketCard({
           >
             Spot Market
           </Typography>
-          <Typography variant="body2" color="text.secondary">
+          <Typography component={`div`} variant="body2" color="text.secondary">
             <Box
               display="flex"
               justifyContent="space-between"