+++ src/web/index.tsx
@@ -1,7 +1,12 @@
 import React from "react";
 import App from "./App";
 import { createRoot } from "react-dom/client";
+import { TerminalContextProvider } from "react-terminal";
 
 const app = document.getElementById(`app`);
 const root = createRoot(app);
-root.render(<App />);
+root.render(
+  <TerminalContextProvider>
+    <App />
+  </TerminalContextProvider>,
+);